+++ Demo/Format-High.test.ps1
@@ -21,33 +21,21 @@ task Examples {
 }
 
 task UnknownArguments {
-	$ErrorActionPreference = $err = 'Continue'
-	try {
-		1 | Format-High.ps1 -Unknown foo
-	}
-	catch { $err = $_ | Out-String }
-	$err
-	assert ($err -like '*\Format-High.ps1 : Unknown arguments: -Unknown foo*\Format-High.test.ps1:*')
+	$ErrorActionPreference = 'Continue'
+	($e = try {1 | Format-High.ps1 -Unknown foo} catch {$_ | Out-String})
+	assert ($e -like '*\Format-High.ps1 : Unknown arguments: -Unknown foo*\Format-High.test.ps1:*')
 }
 
 task BadProperty {
-	$ErrorActionPreference = $err = 'Continue'
-	try {
-		1..11 | Format-High.ps1 {throw 'Oops.'}
-	}
-	catch { $err = $_ | Out-String }
-	$err
-	assert ($err -like '*\Format-High.ps1 : Error on Property evaluation: Oops.*\Format-High.test.ps1:*')
+	$ErrorActionPreference = 'Continue'
+	($e = try {1..11 | Format-High.ps1 {throw 'Oops.'}} catch {$_ | Out-String})
+	assert ($e -like '*\Format-High.ps1 : Error on Property evaluation: Oops.*\Format-High.test.ps1:*')
 }
 
 task BadColorThrow {
-	$ErrorActionPreference = $err = 'Continue'
-	try {
-		1..11 | Format-High.ps1 -Color {throw 'Oops.'}
-	}
-	catch { $err = $_ | Out-String }
-	$err
-	assert ($err -like '*\Format-High.ps1 : Error on Color evaluation: Oops.*\Format-High.test.ps1:*')
+	$ErrorActionPreference = 'Continue'
+	($e = try {1..11 | Format-High.ps1 -Color {throw 'Oops.'}} catch {$_ | Out-String})
+	assert ($e -like '*\Format-High.ps1 : Error on Color evaluation: Oops.*\Format-High.test.ps1:*')
 }
 
 task BadColorKeyButWorks {